+++ app/Http/Controllers/TimbanganController.php
@@ -8,6 +8,7 @@ use App\Models\Blok;
 use App\Models\Hasil;
 use App\Models\HasilHasKaryawan;
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 
 class TimbanganController extends Controller
 {
@@ -40,6 +41,10 @@ class TimbanganController extends Controller
      */
     public function show(Timbangan $timbangan)
     {
+        if (!Auth::user()->can('laporan-list')) {
+            abort(403, 'Anda tidak memiliki hak akses untuk melihat laporan');
+        }
+
         $hasils = Hasil::where('timbangan_id', $timbangan->id)->get();
         $bloks = Blok::get();
         $karyawans = User::role('karyawan')->get();
@@ -52,6 +57,10 @@ class TimbanganController extends Controller
      */
     public function edit(Timbangan $timbangan)
     {
+        if (!Auth::user()->can('laporan-edit')) {
+            abort(403, 'Anda tidak memiliki hak akses untuk mengubah laporan');
+        }
+
         $karyawans = User::role('Karyawan')->get();
         $mandors = User::role('Mandor')->get();
         $bloks = Blok::get();
@@ -65,6 +74,9 @@ class TimbanganController extends Controller
      */
     public function update(Request $request, Timbangan $timbangan)
     {
+        if (!Auth::user()->can('laporan-edit')) {
+            abort(403, 'Anda tidak memiliki hak akses untuk mengubah laporan');
+        }
         // Remove old record by timbangan id
         Hasil::where('timbangan_id', $timbangan->id)->delete();
